@@ -1,33 +1,40 @@
 {
-  "name": "webpack-commons-chunk-plugin-example",
+  "name": "webpack",
   "version": "1.0.0",
+  "scripts": {
+    "dev": "webpack --mode development --watch",
+    "build": "webpack --mode production"
+  },
   "devDependencies": {
-    "babel-core": "^6.25.0",
-    "babel-loader": "^6.4.1",
+    "babel-core": "^6.26.3",
+    "babel-loader": "^7.1.5",
+    "babel-preset-env": "^1.7.0",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-0": "^6.24.1",
-    "eslint": "^3.19.0",
-    "eslint-loader": "^1.7.1",
+    "eslint": "^5.3.0",
+    "eslint-loader": "^2.1.0",
     "gulp": "^3.9.1",
-    "gulp-autoprefixer": "^3.1.1",
+    "gulp-autoprefixer": "^5.0.0",
     "gulp-cached": "^1.1.1",
     "gulp-filter": "^5.1.0",
     "gulp-notify": "^3.0.0",
     "gulp-plumber": "^1.1.0",
-    "gulp-prettier": "^1.1.0",
-    "gulp-prettier-plugin": "^1.2.0",
-    "gulp-replace": "^0.5.4",
-    "gulp-sass": "^3.1.0",
-    "gulp-uglify": "2.1.2",
-    "gulp-watch": "^4.3.11",
-    "prettier": "^1.11.1",
+    "gulp-prettier": "^2.0.0",
+    "gulp-prettier-plugin": "^1.3.0",
+    "gulp-replace": "^1.0.0",
+    "gulp-sass": "^4.0.1",
+    "gulp-uglify": "3.0.1",
+    "gulp-watch": "^5.0.1",
+    "path": "^0.12.7",
+    "prettier": "^1.14.2",
     "uglify-save-license": "^0.4.1",
-    "velocity-animate": "^1.5.1",
-    "webpack": "^2.7.0"
+    "velocity-animate": "^1.5.2",
+    "webpack": "^4.16.5",
+    "webpack-cli": "^3.1.0"
   },
   "dependencies": {
-    "events": "^1.1.1",
+    "events": "^3.0.0",
     "jquery": "^3.2.1",
     "tapjs": "^1.0.9",
     "animejs": "^2.0.2"
