@@ -1,8 +1,5 @@
-var Animation = require('./modules/animation');
-var ClickAnimation = require('./modules/clickAnimation');
+import Animation from './modules/animation';
+import ClickAnimation from './modules/clickAnimation';
 
 var animation = new Animation();
 var clickAnimation = new ClickAnimation();
-
-
-
